add tests for increased coverage
