+++ package.json
@@ -10,8 +10,7 @@
   "scripts": {
     "build": "ethereumjs-config-build",
     "prepublishOnly": "npm run test && npm run build",
-    "coverage": "ethereumjs-config-coverage",
-    "coveralls": "ethereumjs-config-coveralls",
+    "coverage": "npx nyc npm run test:node",
     "docs:build": "typedoc --out docs --mode file --readme none --theme markdown --mdEngine github --excludeNotExported src",
     "format": "ethereumjs-config-format",
     "format:fix": "ethereumjs-config-format-fix",
@@ -20,9 +19,9 @@
     "tsc": "ethereumjs-config-tsc",
     "lint": "ethereumjs-config-lint",
     "lint:fix": "ethereumjs-config-lint-fix",
-    "test": "npm run test:node",
-    "test:node": "ts-node node_modules/tape/bin/tape ./test/index.ts",
-    "test:browser:build": "tsc && cp ./test/*.json test-build/test/",
+    "test": "npm run test:node && npm run test:browser",
+    "test:node": "npx tape -r ts-node/register test/*.spec.ts",
+    "test:browser:build": "tsc && cp test/testdata/*.json test-build/test/testdata",
     "test:browser": "npm run test:browser:build && karma start karma.conf.js"
   },
   "husky": {
@@ -56,18 +55,19 @@
     "@ethereumjs/config-prettier": "^1.1.1",
     "@ethereumjs/config-tsc": "^1.1.1",
     "@ethereumjs/config-tslint": "^1.1.1",
+    "@types/lru-cache": "^5.1.0",
     "@types/node": "^11.13.4",
     "@types/tape": "^4.2.33",
     "browserify": "^16.2.3",
-    "coveralls": "^2.11.4",
+    "ethereumjs-blockchain": "^4.0.3",
     "husky": "^2.1.0",
     "istanbul": "^0.4.1",
     "karma": "^4.1.0",
     "karma-browserify": "^6.0.0",
-    "karma-detect-browsers": "^2.3.3",
+    "karma-chrome-launcher": "^2.0.0",
     "karma-firefox-launcher": "^1.1.0",
     "karma-tap": "^4.1.4",
-    "nyc": "^14.0.0",
+    "nyc": "^15.0.0",
     "prettier": "^1.17.0",
     "tape": "^4.0.3",
     "ts-node": "^8.0.3",
