--- test/index.ts
@@ -1,4 +1,0 @@
-import './header'
-import './block'
-import './difficulty'
-import './from-rpc'
