+++ test/testdata/difficultyFrontier.json
