+++ test/testdata/difficultyMainNetwork.json
